refactor (Constants): 상수 구조 개선
