+++ src/main/kotlin/omok/configure/Constants.kt
@@ -1,0 +1,8 @@
+package omok.configure
+
+object Constants {
+    const val X_MIN_RANGE = 1
+    const val X_MAX_RANGE = 15
+    const val Y_MIN_RANGE = 1
+    const val Y_MAX_RANGE = 15
+}
